% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRAPE_Basic_Functions.R
\name{makePairwiseOrder}
\alias{makePairwiseOrder}
\title{Make pairwise order representation of a sample}
\usage{
makePairwiseOrder(samp)
}
\arguments{
\item{samp}{A vector of gene expression values}
}
\value{
Binary vector of the pairwise ranking representation of the samples
}
\description{
Takes in a vector of gene expression values and returns a binary vector consisting of the pairwise rankings for the sample
}
\examples{
samp <- c(1,3,2,1.5)
makePairwiseOrder(samp)
}
