% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRAPE_Classification_Functions.R
\name{w_quad}
\alias{w_quad}
\title{Quadratic weight function}
\usage{
w_quad(x)
}
\arguments{
\item{x}{Any number, vector of matrix.}
}
\value{
Weight of each element
}
\description{
Calculates the weights of all input entries. All entries should take values in [0,1].
}
\examples{
w_quad(0.95)
w_quad(cbind(c(.7,.8),c(.9,.1)))
}
